+++ trunk/src/by/bsu/mg/math/utils/Point2d.java
@@ -41,18 +41,18 @@ public class Point2d implements IPoint{
         y = coords[1];
     }
 
-    public double[] to2dArray(){
-        double[] pointArr = new double[2];
-        pointArr[0] = x;
-        pointArr[1] = y;
+    public float[] to2dArray(){
+        float[] pointArr = new float[2];
+        pointArr[0] = (float) x;
+        pointArr[1] = (float) y;
 
         return pointArr;
     }
 
-    public double[] to3dArray(){
-        double[] pointArr = new double[3];
-        pointArr[0] = x;
-        pointArr[1] = y;
+    public float[] to3dArray(){
+        float[] pointArr = new float[3];
+        pointArr[0] = (float) x;
+        pointArr[1] = (float) y;
         pointArr[2] = 0;
 
         return pointArr;
